base

+++ src/main/java/com/Configure/SwaggerApi/repository/ProductRepository.java
@@ -3,5 +3,7 @@ package com.Configure.SwaggerApi.repository;
 import com.Configure.SwaggerApi.entitiy.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductRepository extends JpaRepository<Product, Integer> {
+import java.util.List;
+
+public interface ProductRepository extends JpaRepository<Product, Long> {
 }
